requirements.txt
numpy
pandas
matplotlib
seaborn
imageio
pillow
scikit-learn
